+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This page shows a detailed overview of the changes between versions without the 
 
 ## __WORK IN PROGRESS__
 
+- Enhancement: Introduce /health endpoint to use for (docker) health checks. It returns the server version and number of nodes
 - Adjustment: Adhere to the default nodeId 112233 for the controller itself as the Python Matter server did
 - Adjustment: For fresh starts the next node id to be commissioned will be 1 as it was for the Python Matter Server
 - Fix: Start up the server also when no nodes exist in the migrated json file. Logs a warning